# swagger/swagger.yaml
definitions:
...
  AuthResponse:
    description: Response containing JWT token and user details after login.
    properties:
      refreshToken:
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/User'
    required:
      - token
      - user
    type: object
...
  ErrorResponse:
    description: Standard error response.
    properties:
      error:
        example: BadRequest
        type: string
      message:
        example: Invalid request parameters
        type: string
    required:
      - error
      - message
    type: object
...
  LoginRequest:
    description: Payload to authenticate a user.
    properties:
      email:
        example: paras@example.com
        format: email
        type: string
      password:
        example: strongPassword123
        type: string
    required:
      - email
      - password
    type: object
...
  RefreshTokenRequest:
    description: Payload to refresh authentication token.
    properties:
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
      - refreshToken
    type: object
...
  RegisterRequest:
    description: Payload to register a new user.
    properties:
      email:
        example: paras@example.com
        format: email
        type: string
      name:
        example: Paras Jain
        type: string
      password:
        example: strongPassword123
        type: string
      phone:
        example: 919876543210
        type: string
    required:
      - name
      - email
      - password
    type: object
...
  SuccessResponse:
    description: Standard success response.
    properties:
      message:
        example: Operation successful
        type: string
    required:
      - message
    type: object
...
  User:
    description: Represents a registered user in the system.
    properties:
      createdAt:
        description: User creation timestamp
        format: date-time
        type: string
      email:
        description: User email address
        example: paras@example.com
        format: email
        type: string
      id:
        description: Unique user identifier
        example: 101
        type: integer
      name:
        description: Full name of the user
        example: Paras Jain
        type: string
      password:
        description: Password of user
        type: string
      phone:
        description: User phone number
        example: 919876543210
        type: string
      updatedAt:
        description: Last update timestamp
        format: date-time
        type: string
    required:
      - id
      - email
      - name
    type: object
...
  UserUpdateRequest:
    description: Payload to update user profile.
    properties:
      name:
        example: Paras Jain
        type: string
      phone:
        example: 919876543210
        type: string
    type: object
info:
  title: Adornme
  version: 1.0.0
paths:
  /auth/forgot-password:
    post:
      consumes:
        - application/json
      description: Sends a reset password link to the user's email
      operationId: forgetPassword
      parameters:
        - in: body
          name: body
          required: true
          schema:
            properties:
              email:
                example: user@example.com
                type: string
            required:
              - email
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: Reset link sent (if email exists)
          schema:
            properties:
              message:
                example: If the email exists, a reset link has been sent
                type: string
            type: object
        "400":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Request password reset
      tags:
        - Users
  /auth/login:
    post:
      consumes:
        - application/json
      operationId: loginUser
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      produces:
        - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/AuthResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Login user
      tags:
        - Users
  /auth/logout:
    post:
      operationId: logoutUser
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Logout user
      tags:
        - Users
  /auth/refresh-token:
    post:
      consumes:
        - application/json
      operationId: refreshToken
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/RefreshTokenRequest'
      produces:
        - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/AuthResponse'
        "400":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Refresh authentication token
      tags:
        - Users
  /auth/register:
    post:
      consumes:
        - application/json
      operationId: registerUser
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/RegisterRequest'
      produces:
        - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/AuthResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Register a new user
      tags:
        - Users
...
  /users/me:
    get:
      operationId: getUserProfile
      produces:
        - application/json
      responses:
        "200":
          description: User profile details
          schema:
            $ref: '#/definitions/User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Get logged-in user profile
      tags:
        - Users
    put:
      consumes:
        - application/json
      operationId: updateUserProfile
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UserUpdateRequest'
      produces:
        - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Update logged-in user profile
      tags:
        - Users
  /users/{id}:
    delete:
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        "204":
          description: User deleted
      security:
        - bearerAuth: []
      summary: Delete a user
      tags:
        - AdminUsers
    get:
      operationId: getUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: User details
      security:
        - bearerAuth: []
      summary: Get user details
      tags:
        - AdminUsers
    put:
      operationId: updateUser
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UserUpdateRequest'
      responses:
        "200":
          description: User updated
      security:
        - bearerAuth: []
      summary: Update user info
      tags:
        - AdminUsers
schemes:
  - http
securityDefinitions:
  apiKey:
    in: header
    name: x-api-key
    type: apiKey
  basicAuth:
    type: basic
  bearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"